Move logic to mod.rs
